+++ CHANGES.rst
@@ -109,3 +109,4 @@ v1.0.16[unreleased]
 * Deprecate /country-cities/ route and implement /available-cities/ route
 * Fix UserSearchResource name/autocompletion filtering
 * Fix haystack filter when using ElasticSearch as backend
+* Allow filtering a component by multiple types using an OR filter. (Pass multiple components with different types to use AND filter)